Hide NullDimensionSelector from public
