+++ processing/src/main/java/org/apache/druid/segment/ConstantDimensionSelector.java
@@ -35,11 +35,11 @@ public class ConstantDimensionSelector implements SingleValueHistoricalDimension
 {
   private final String value;
 
-  public ConstantDimensionSelector(final String value)
+  ConstantDimensionSelector(final String value)
   {
     if (NullHandling.isNullOrEquivalent(value)) {
       // There's an optimized implementation for nulls that callers should use instead.
-      throw new IllegalArgumentException("Use NullDimensionSelector or DimensionSelectorUtils.constantSelector");
+      throw new IllegalArgumentException("Use DimensionSelector.constant(null)");
     }
 
     this.value = value;
